Working code
I got the tracking to work (I think).

This contains an example that will run, but if you want to see it in
action you'll have to modify the code to calibrate it.  The method for
calibration already exist, but there is no code referencing it.

I was able to run it on a Raspberry PI and get ~5FPS so I think things
are looking pretty good.  Currently it doesn't do much beside find the
center of the object being tracked, but those values are not exposed,
and nothing is done with it following processing.